extract nmcli helpers
